+++ backend/cleanpro/.dockerignore
@@ -1,0 +1,6 @@
+venv
+.git
+db.sqlite3
+.idea
+.vscode
+.env
