+++ backend/cleanpro/requirements.txt
@@ -7,14 +7,17 @@ defusedxml==0.7.1
 Django==4.2.4
 django-filter==23.2
 django-password-validators==1.7.1
+django-phone-field==1.8.1
 django-phonenumber-field==7.1.0
 django-phonenumbers==1.0.1
 django-templated-mail==1.1.1
 djangorestframework==3.14.0
 djangorestframework-simplejwt==5.3.0
 djoser==2.2.0
+drf-yasg==1.21.7
 gunicorn==21.2.0
 idna==3.4
+inflection==0.5.1
 oauthlib==3.2.2
 packaging==23.1
 phonenumbers==8.13.19
@@ -25,11 +28,13 @@ PyJWT==2.8.0
 python-dotenv==1.0.0
 python3-openid==3.2.0
 pytz==2023.3
+PyYAML==6.0.1
 requests==2.31.0
 requests-oauthlib==1.3.1
 social-auth-app-django==5.2.0
 social-auth-core==4.4.2
 sqlparse==0.4.4
+typing_extensions==4.7.1
 tzdata==2023.3
+uritemplate==4.1.1
 urllib3==2.0.4
-django-phone-field==1.8.1
