+++ docker-compose.yml
@@ -1,14 +1,37 @@
 version: '3.8'
 
-services:
-  db:
-    image: postgres
-    ports:
-      - ${DB_PORT}:${DB_PORT}
-    volumes:
-      - database:/var/lib/postgresql/data/
-    env_file:
-      - ./.env
-
 volumes:
-  database:
+  database_cs:
+  static_cs:
+  media_cs:
+
+services:
+  db_cs:
+    image: postgres
+    volumes:
+      - database_cs:/var/lib/postgresql/data/
+    env_file:
+      - backend/cleanpro/.env
+  backend:
+    build:
+      context: backend/cleanpro
+      dockerfile: Dockerfile
+    env_file: backend/cleanpro/.env
+    volumes:
+      - static_cs:/static
+      - media_cs:/app/media
+    depends_on:
+      - db_cs
+  nginx:
+    image: nginx:1.19.3
+    ports:
+      - "80:80"
+    volumes:
+      - ./nginx.conf:/etc/nginx/conf.d/default.conf
+      # - ../frontend/build:/usr/share/nginx/html/
+      - ../docs/:/usr/share/nginx/html/docs/
+      - static_cs:/var/html/static/
+      - media_cs:/var/html/media/
+    depends_on:
+      - backend
+      # - frontend
