+++ nginx.conf
@@ -1,0 +1,46 @@
+server {
+    listen 80;
+    server_tokens off;
+
+    location /docs/ {
+        root /usr/share/nginx/html;
+        try_files $uri $uri/redoc.html;
+    }
+
+    location / {
+        proxy_set_header Host $http_host;
+        proxy_set_header        X-Real-IP $remote_addr;
+        proxy_set_header        X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_pass http://backend:8000/;
+    }
+
+    location /admin/ {
+        proxy_set_header Host $http_host;
+        proxy_pass http://backend:8000/admin/;
+    }
+  
+    location /media/ {
+        root /var/html/;
+    }
+    location /static/admin/ {
+        root /var/html/;
+    }
+    location /static/rest_framework/ {
+        root /var/html/;
+    }
+
+    # location / {
+    #     root /usr/share/nginx/html;
+    #     index  index.html index.htm;
+    #     try_files $uri /index.html;
+    #     proxy_set_header Host $http_host;
+    #     proxy_set_header        X-Real-IP $remote_addr;
+    #     proxy_set_header        X-Forwarded-For $proxy_add_x_forwarded_for;
+    #     proxy_set_header        X-Forwarded-Proto $scheme;
+    #   }
+    #   error_page   500 502 503 504  /50x.html;
+    #   location = /50x.html {
+    #     root   /var/html/frontend/;
+    #   }
+
+}
